refactor: 댓글 리스트 컴포넌트로 분리
